requirements.txt
streamlit==1.31.0
mistralai==1.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
```
langchain==0.0.208
chromadb==0.3.26    